requirements.txt
openai==1.101.0
pymupdf==1.26.4
PyPDF2==3.0.1
python-dotenv==1.1.0
tiktoken==0.11.0
pyyaml==6.0.2
fastapi==0.118.2
uvicorn==0.37.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic-settings==2.11.0
python-multipart==0.0.20
pgvector==0.3.6
